+++ RELEASE_NOTES.md
@@ -1,7 +1,9 @@
 ### 0.2-dev - Unreleased
-* Switch to using Milekic.YoLo
+* Switch to using YoLo gist
 * PublicKeyEncryption.encrypt now returns nonce in addition to cipher text to simplify interface usage
 * Add Random
+* Add PasswordHasing
+* Add Secret
 
 ### 0.1.1 - 2018/04/25
 * Move YoLo to FsSodium namespace